[Refactor] Move functional out of experimental
