+++ include/occa.hpp
@@ -3,6 +3,7 @@
 
 #include <occa/core.hpp>
 #include <occa/dtype.hpp>
+#include <occa/functional.hpp>
 #include <occa/types.hpp>
 #include <occa/utils.hpp>
