+++ include/occa/c/experimental.h
@@ -2,7 +2,6 @@
 #define OCCA_C_EXPERIMENTAL_HEADER
 
 #include <occa/c/experimental/kernelBuilder.h>
-#include <occa/c/experimental/scope.h>
 
 #ifdef OCCA_JIT
 #  undef OCCA_JIT
