+++ include/occa/c/experimental/kernelBuilder.h
@@ -3,7 +3,7 @@
 
 #include <occa/c/defines.h>
 #include <occa/c/types.h>
-#include <occa/c/experimental/scope.h>
+#include <occa/c/scope.h>
 
 OCCA_START_EXTERN_C
