+++ include/occa/c/scope.h
