+++ include/occa/defines/okl.hpp
@@ -1,7 +1,9 @@
-// Just in case someone wants to run with an older format than C99
-#ifndef OCCA_DISABLE_VARIADIC_MACROS
-#  ifndef OCCA_DEFINES_OKL_HEADER
-#  define OCCA_DEFINES_OKL_HEADER
+#ifndef OCCA_DEFINES_OKL_HEADER
+#define OCCA_DEFINES_OKL_HEADER
+
+#ifdef OCCA_JIT
+#  undef OCCA_JIT
+#endif
 
 #define OCCA_JIT(OKL_SCOPE, OKL_SOURCE)                 \
   do {                                                  \
@@ -12,5 +14,4 @@
     _occaJitKernelBuilder.run(OKL_SCOPE);               \
   } while (false)
 
-#  endif
 #endif
