+++ include/occa/experimental.hpp
@@ -1,7 +1,6 @@
 #ifndef OCCA_EXPERIMENTAL_HEADER
 #define OCCA_EXPERIMENTAL_HEADER
 
-#include <occa/experimental/functional.hpp>
 #include <occa/experimental/kernelBuilder.hpp>
 #include <occa/experimental/mpi.hpp>
