--- include/occa/experimental/functional.hpp
@@ -1,9 +1,0 @@
-#ifndef OCCA_EXPERIMENTAL_FUNCTIONAL_HEADER
-#define OCCA_EXPERIMENTAL_FUNCTIONAL_HEADER
-
-#include <occa/experimental/functional/array.hpp>
-#include <occa/experimental/functional/function.hpp>
-#include <occa/experimental/functional/scope.hpp>
-#include <occa/experimental/functional/utils.hpp>
-
-#endif
