+++ include/occa/functional.hpp
@@ -1,0 +1,9 @@
+#ifndef OCCA_FUNCTIONAL_HEADER
+#define OCCA_FUNCTIONAL_HEADER
+
+#include <occa/functional/array.hpp>
+#include <occa/functional/function.hpp>
+#include <occa/functional/scope.hpp>
+#include <occa/functional/utils.hpp>
+
+#endif
