+++ include/occa/functional/scope.hpp
@@ -1,5 +1,5 @@
-#ifndef OCCA_EXPERIMENTAL_FUNCTIONAL_SCOPE_HEADER
-#define OCCA_EXPERIMENTAL_FUNCTIONAL_SCOPE_HEADER
+#ifndef OCCA_FUNCTIONAL_SCOPE_HEADER
+#define OCCA_FUNCTIONAL_SCOPE_HEADER
 
 #include <initializer_list>
 #include <map>
