+++ src/occa/internal/c/types.hpp
@@ -2,7 +2,7 @@
 #define OCCA_INTERNAL_C_TYPES_HEADER
 
 #include <occa/core.hpp>
-#include <occa/experimental/functional/scope.hpp>
+#include <occa/functional/scope.hpp>
 #include <occa/experimental/kernelBuilder.hpp>
 #include <occa/internal/utils/sys.hpp>
